Split code into modules
 * module eQ3 is the connector to the eQ-3 thermostat
 * module interface is the interface between the thermostat and the
   valve
 * module RAV is the connector to the valve

interface_h is an important number. On my valves this was adjusted to
match the distance between the eQ-3 motor and the valve pin. A value
smaller than 1mm would mean that the eQ-3 would not be capable of fully
opening the valve.

Also there should be a gap of 3.2 mm between the interface and the dent
on the eQ-3 side for hosting the eQ-3 clips.

I have four instances of this part running at my place but while it is
working, I found some issues:
 * The RAV connector barely stays in place. Maybe the torus is in the
wrong position or has the wrong dimentions?
 * The eQ-3 connector needs deeper grooves for the clips for a better
grip.

Signed-off-by: Peter Senna Tschudin <peter.senna@gmail.com>